# packages/observing-shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Observ.ing API",
    "version": "0.1.0-alpha",
    "description": "**⚠️ This API is unstable and subject to change without notice.**\n\nDecentralized biodiversity observation platform API built on AT Protocol."
  },
  "servers": [
    {
      "url": "https://biosky-appview-378013734384.us-central1.run.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "schemas": {
      "Profile": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string",
            "description": "Decentralized identifier"
          },
          "handle": {
            "type": "string",
            "description": "User handle"
          },
          "displayName": {
            "type": "string",
            "description": "Display name"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar URL"
          }
        },
        "required": [
          "did"
        ]
      },
      "Observer": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Profile"
          },
          {
            "type": "object",
            "properties": {
              "role": {
                "type": "string",
                "enum": [
                  "owner",
                  "co-observer"
                ],
                "description": "Observer role"
              }
            },
            "required": [
              "role"
            ]
          }
        ]
      },
      "Subject": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0,
            "description": "Subject index within the observation"
          },
          "communityId": {
            "type": "string",
            "description": "Community-agreed taxon name"
          },
          "identificationCount": {
            "type": "integer",
            "description": "Number of identifications"
          }
        },
        "required": [
          "index",
          "identificationCount"
        ]
      },
      "Location": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "description": "Decimal latitude"
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "description": "Decimal longitude"
          },
          "uncertaintyMeters": {
            "type": "number",
            "description": "Coordinate uncertainty in meters"
          },
          "continent": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "stateProvince": {
            "type": "string"
          },
          "county": {
            "type": "string"
          },
          "municipality": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "waterBody": {
            "type": "string"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "Occurrence": {
        "type": "object",
        "properties": {
          "uri": {
            "type": "string",
            "description": "AT Protocol URI"
          },
          "cid": {
            "type": "string",
            "description": "Content identifier"
          },
          "observer": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Profile"
              },
              {
                "description": "Primary observer"
              }
            ]
          },
          "observers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Observer"
            },
            "description": "All observers including co-observers"
          },
          "scientificName": {
            "type": "string",
            "description": "[DEPRECATED] Scientific name from record - use effectiveTaxonomy instead"
          },
          "communityId": {
            "type": "string",
            "description": "Community ID for subject 0 (backward compat)"
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Subject"
            },
            "description": "All subjects with their community IDs"
          },
          "eventDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date of observation"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "verbatimLocality": {
            "type": "string"
          },
          "occurrenceRemarks": {
            "type": "string"
          },
          "taxonId": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "taxonRank": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "vernacularName": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "kingdom": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "phylum": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "class": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "order": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "family": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "genus": {
            "type": "string",
            "description": "[DEPRECATED] Use effectiveTaxonomy instead"
          },
          "effectiveTaxonomy": {
            "$ref": "#/components/schemas/EffectiveTaxonomy"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Image URLs"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "identificationCount": {
            "type": "integer"
          }
        },
        "required": [
          "uri",
          "cid",
          "observer",
          "observers",
          "subjects",
          "eventDate",
          "location",
          "images",
          "createdAt"
        ]
      },
...
      "CreateOccurrenceRequest": {
        "type": "object",
        "properties": {
          "scientificName": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "notes": {
            "type": "string"
          },
          "license": {
            "type": "string"
          },
          "eventDate": {
            "type": "string",
            "format": "date-time"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "description": "Base64 encoded image data"
                },
                "mimeType": {
                  "type": "string",
                  "description": "MIME type of the image"
                }
              },
              "required": [
                "data",
                "mimeType"
              ]
            }
          },
          "taxonId": {
            "type": "string"
          },
          "taxonRank": {
            "type": "string"
          },
          "vernacularName": {
            "type": "string"
          },
          "kingdom": {
            "type": "string"
          },
          "phylum": {
            "type": "string"
          },
          "class": {
            "type": "string"
          },
          "order": {
            "type": "string"
          },
          "family": {
            "type": "string"
          },
          "genus": {
            "type": "string"
          },
          "recordedBy": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Co-observer DIDs"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "CreateOccurrenceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "uri": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "uri",
          "cid"
        ]
      },
      "Identification": {
        "type": "object",
        "properties": {
          "uri": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          },
          "did": {
            "type": "string"
          },
          "subject_uri": {
            "type": "string"
          },
          "subject_index": {
            "type": "integer"
          },
          "scientific_name": {
            "type": "string"
          },
          "taxon_rank": {
            "type": "string"
          },
          "identification_remarks": {
            "type": "string"
          },
          "is_agreement": {
            "type": "boolean"
          },
          "date_identified": {
            "type": "string",
            "format": "date-time"
          },
          "identifier": {
            "$ref": "#/components/schemas/Profile"
          },
          "kingdom": {
            "type": "string"
          }
        },
        "required": [
          "uri",
          "cid",
          "did",
          "subject_uri",
          "subject_index",
          "scientific_name",
          "is_agreement",
          "date_identified"
        ]
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "uri": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          },
          "did": {
            "type": "string"
          },
          "subject_uri": {
            "type": "string"
          },
          "subject_cid": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "reply_to_uri": {
            "type": "string"
          },
          "reply_to_cid": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "commenter": {
            "$ref": "#/components/schemas/Profile"
          }
        },
        "required": [
          "uri",
          "cid",
          "did",
          "subject_uri",
          "subject_cid",
          "body",
          "created_at"
        ]
      },
...
      "FeedResponse": {
        "type": "object",
        "properties": {
          "occurrences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Occurrence"
            }
          },
          "cursor": {
            "type": "string",
            "description": "Cursor for pagination"
          }
        },
        "required": [
          "occurrences"
        ]
      },
...
      "GeoJSONFeature": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "Feature"
            ]
          },
          "geometry": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Point"
                ]
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "number"
                },
                "minItems": 2,
                "maxItems": 2
              }
            },
            "required": [
              "type",
              "coordinates"
            ]
          },
          "properties": {
            "type": "object",
            "properties": {
              "uri": {
                "type": "string"
              },
              "scientificName": {
                "type": "string"
              },
              "eventDate": {
                "type": "string"
              },
              "point_count": {
                "type": "integer"
              },
              "cluster_id": {
                "type": "integer"
              }
            },
            "required": [
              "uri"
            ]
          }
        },
        "required": [
          "type",
          "geometry",
          "properties"
        ]
      },
      "GeoJSONFeatureCollection": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "FeatureCollection"
            ]
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            }
          }
        },
        "required": [
          "type",
          "features"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "OccurrenceDetailResponse": {
        "type": "object",
        "properties": {
          "occurrence": {
            "$ref": "#/components/schemas/Occurrence"
          },
          "identifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Identification"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          }
        },
        "required": [
          "occurrence",
          "identifications",
          "comments"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/occurrences/feed": {
      "get": {
        "tags": [
          "Occurrences"
        ],
        "summary": "Get chronological occurrence feed",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Feed of occurrences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/occurrences/nearby": {
      "get": {
        "tags": [
          "Occurrences"
        ],
        "summary": "Get occurrences near a location",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "lat",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "lng",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "radius",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Nearby occurrences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/occurrences/geojson": {
      "get": {
        "tags": [
          "Occurrences"
        ],
        "summary": "Get occurrences as GeoJSON for mapping",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "minLat",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "minLng",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "maxLat",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "maxLng",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "GeoJSON feature collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          }
        }
      }
    },
    "/api/occurrences/{uri}": {
      "get": {
        "tags": [
          "Occurrences"
        ],
        "summary": "Get single occurrence with identifications and comments",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "uri",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Occurrence detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Occurrence not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/occurrences": {
      "post": {
        "tags": [
          "Occurrences"
        ],
        "summary": "Create a new occurrence",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOccurrenceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Occurrence created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOccurrenceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/taxa/{id}/occurrences": {
      "get": {
        "tags": [
          "Taxonomy"
        ],
        "summary": "Get occurrences for a taxon",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Taxon occurrences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}